Adding all files from the Pong Game
This pong game was created in the Unity Pong Game section of the Complete C# Masterclass course from Udemy. It was not added to a repository sooner because I had problems attempting to add these files using git. Files were committed using Github Desktop. All future changes to this project will be commited to this repository.